+++ e2e-wdio/page-objects/index.ts
@@ -1,0 +1,2 @@
+export * from "./login.po";
+export * from "./main.po";
