+++ e2e-wdio/services/index.ts
@@ -1,0 +1,2 @@
+export * from "./login-action.service";
+export * from "./login-navigation.service";
