+++ e2e-wdio/services/login-navigation.service.ts
@@ -1,0 +1,10 @@
+import { LoginPage } from "@e2e-pages";
+import { LoginActionService } from "@e2e-services";
+
+
+export class LoginNavigationService {
+    public static async navigateToHomePage(loginPage: LoginPage) {
+        await browser.url("/#/login");
+        await LoginActionService.setAppParams(loginPage);
+    }
+}
