+++ e2e-wdio/utils/general-parameters.ts
@@ -1,8 +1,8 @@
 export class GeneralParameters {
-	public static readonly NOT_RETRIEVED = '<not retrieved yet, available after calling navigateToHomePage()>';
+	public static readonly NOT_RETRIEVED = "<not retrieved yet, available after calling navigateToHomePage()>";
 
-	public static readonly USERNAME = 'Systelab';
-	public static readonly PASSWORD = 'Systelab';
+	public static readonly USERNAME = "Systelab";
+	public static readonly PASSWORD = "Systelab";
 
 	public static appName = GeneralParameters.NOT_RETRIEVED;
 	public static appDescription = GeneralParameters.NOT_RETRIEVED;
