+++ e2e-wdio/utils/index.ts
@@ -1,0 +1,1 @@
+export * from "./general-parameters";
