+++ package.json
@@ -90,6 +90,7 @@
     "systelab-schematics": "^1.2.10",
     "ts-node": "10.4.0",
     "tslint": "~5.18.0",
+    "tsconfig-paths": "3.8.0",
     "typescript": "^4.2.3",
     "wdio-chromedriver-service": "7.2.2"
   }
