+++ tsconfig.json
@@ -23,7 +23,10 @@
       "@components/*": ["src/core/components/*"],
       "@api/*": ["src/core/api/*"],
       "@model/*": ["src/core/model/*"],
-      "@features/*": ["src/features/*"]
+      "@features/*": ["src/features/*"],
+      "@e2e-pages": ["e2e-wdio/page-objects/index.ts"],
+      "@e2e-services": ["e2e-wdio/services/index.ts"],
+      "@e2e-utils": ["e2e-wdio/utils/index.ts"],
     }
   },
   "angularCompilerOptions": {
